[DB] sail artisan migrate:refresh --path=/database/migrations/product_table.php -> 법정동 코드 BigInteger 로 수정

+++ database/migrations/product_table.php
@@ -24,7 +24,7 @@ return new class extends Migration
             $table->integer('is_map')->nullable()->comment('지도 노출 여부 - 0: 노출, 1: 노출안함');
             $table->string('address_lat')->nullable()->comment('위도');
             $table->string('address_lng')->nullable()->comment('경도');
-            $table->integer('region_code')->nullable()->comment('주소 법정동 코드');
+            $table->bigInteger('region_code')->nullable()->comment('주소 법정동 코드');
             $table->string('region_address')->nullable()->comment('주소 법정동 주소');
             $table->string('address')->nullable()->comment('주소');
             $table->string('address_detail')->nullable()->comment('상세 주소');
